Updated Chapter 20

+++ full/chapters/20-file-io/programs/BaseballPlayer.java
@@ -3,9 +3,9 @@ import java.io.Serializable;
 public class BaseballPlayer implements Serializable {
     private int homeRuns;
     private int RBI;
-    private float battingAverage;
+    private double battingAverage;
     
-    public BaseballPlayer( int homeRuns, int RBI, float battingAverage ) {
+    public BaseballPlayer(int homeRuns, int RBI, double battingAverage) {
         this.homeRuns = homeRuns;
         this.RBI = RBI;
         this.battingAverage = battingAverage;
@@ -13,5 +13,5 @@ public class BaseballPlayer implements Serializable {
     
     public int getHomeRuns() { return homeRuns; } 
     public int getRBI() { return RBI; }
-    public float getBattingAverage() { return battingAverage; }
+    public double getBattingAverage() { return battingAverage; }
 }
